requirements.txt
# Python 3.10 recommended


numpy==1.24.4
opencv-python==4.6.0.66
pillow==9.5.0
requests==2.32.3
regex==2024.11.6

paddlepaddle==2.5.2
paddleocr==2.6.1.3
PyMuPDF==1.20.2

fasttext-wheel==0.9.2

transformers==4.53.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1

clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
ftfy==6.3.1
tqdm==4.67.1

ollama==0.5.1
tiktoken==0.9.0

Flask==3.1.1
